[data] deflake sql + consumption + execution_optimizer
