+++ python/ray/data/tests/test_raydp.py
@@ -4,7 +4,7 @@ import raydp
 
 import ray
 from ray.data.tests.conftest import *  # noqa
-from ray.data.tests.test_execution_optimizer import _check_usage_record
+from ray.data.tests.test_util import _check_usage_record
 
 
 # RayDP tests require Ray Java. Make sure ray jar is built before running this test.
